[CS-3103]: Split component/business logic on assetList
